Keep decompositions, lowering, customops in _inductor
These files are core compiler internals (graph rewrites, LoopLevelIR
lowering, custom op definitions for compilation). They belong in
_inductor/ alongside their consumers and dependencies.

The ops/ package now only contains runtime files: eager.py and
fallbacks.py.
