+++ docs/compiler_architecture.md
@@ -44,8 +44,8 @@ orchestrate their execution.
 
 Some key entry points to the front-end compiler are:
 + [init.py](../torch_spyre/_inductor/__init__.py) registers the compiler and customizes the configuration of Inductor.
-+ [decompositions.py](../torch_spyre/ops/decompositions.py) is where we add Spyre-specific decompositions of existing high-level ATen operations.
-+ [custom_ops.py](../torch_spyre/ops/customops.py) is where we define new Spyre-specific operations.
++ [decompositions.py](../torch_spyre/_inductor/decompositions.py) is where we add Spyre-specific decompositions of existing high-level ATen operations.
++ [custom_ops.py](../torch_spyre/_inductor/customops.py) is where we define new Spyre-specific operations.
 + [passes.py](../torch_spyre/_inductor/passes.py) is where we add Spyre-specific compiler passes into the three exported
 extension points of Inductor. It supports adding passes to both the FX Graph and LoopLevelIR stages of compilation.
 + [spyre_kernel.py](../torch_spyre/_inductor/spyre_kernel.py) defines our compilation from LoopLevelIR into `OpSpec`, our
