+++ torch_spyre/_inductor/__init__.py
@@ -121,7 +121,7 @@ def enable_spyre_compile_fx_wrapper():
 
 
 def _light_autoload():
-    from torch_spyre.ops import decompositions  # noqa: F401
+    from . import decompositions  # noqa: F401
 
     enable_spyre_compile_fx_wrapper()
